Refactored test_ship_sunk_destroyer_player into test_destroyer file. Also refactored ship_sunk_destroyer_player into destroyer.py.
